Get rid of compiler warning.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@14751 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/RevisionGraph/StandardLayoutTextList.cpp
@@ -105,7 +105,7 @@ CStandardLayoutTextList::GetText (index_t index) const
         if ((textInfo.subPathIndex == 1) && (nodeInfo.skipStartPathElements > 0))
             text.Insert (0, CString ('.', nodeInfo.skipStartPathElements));
 
-        if (   (visibleElementCount == textInfo.subPathIndex) 
+        if (   (visibleElementCount == (size_t)textInfo.subPathIndex) 
             && (nodeInfo.skipTailPathElements != 0))
         {
             text.AppendChar (_T('/'));
